stash view redesigned
