+++ core/git/file.go
@@ -20,6 +20,8 @@ var (
 	StatusNotupdated FileStatus = ' '
 	// StatusModified says file is modifed
 	StatusModified FileStatus = 'M'
+	// StatusModifiedUntracked says file is modifed and untracked
+	StatusModifiedUntracked FileStatus = 'm'
 	// StatusAdded says file is added to index
 	StatusAdded FileStatus = 'A'
 	// StatusDeleted says file is deleted
