+++ gui/overview.go
@@ -47,11 +47,11 @@ func (gui *Gui) overviewLayout(g *gocui.Gui) error {
 		v.Wrap = false
 		v.Autoscroll = false
 	}
-	if v, err := g.SetView(stashedViewFeature.Name, -1*int(0.20*float32(maxX)), 0, -1, maxY); err != nil {
+	if v, err := g.SetView(stashViewFeature.Name, -1*int(0.20*float32(maxX)), 0, -1, maxY); err != nil {
 		if err != gocui.ErrUnknownView {
 			return err
 		}
-		v.Title = stashedViewFeature.Title
+		v.Title = stashViewFeature.Title
 		v.Wrap = false
 		v.Autoscroll = false
 	}
